[fix] Fix list saving and message corruption
The list was shown with incorrect values. This was fixed updating the
functions with strdup() instead of attributing the pointer. After that
was fixed, the messages started displaying corruption on the data. That
was also fixed by updating the functions of sending and receiving
messages to send() and recv() instead of write() and read().